requirements.txt
matplotlib==3.5.2
numpy==1.23.1
Pillow==9.4.0
PySimpleGUI==4.60.4
python_pptx==0.6.21
scikit_image==0.19.3
scipy==1.9.0
seaborn==0.11.2
skimage==0.0
